Moved service popup component, small css things
