+++ src/components/Service.tsx
@@ -1,3 +1,5 @@
+import { Properties } from 'csstype';
+import { ChangeEvent, MouseEvent, useState } from 'react';
 import { AppState, defaultEmptyApp, Dispatch } from '../state';
 
 type ServiceProps = {
@@ -40,6 +42,61 @@ export const Service = (Props: ServiceProps) => {
     return (
         <div id={`service-${service.id}`} data-service={JSON.stringify(service)} onClick={clickService}>
             {name}:{port}
+        </div>
+    );
+};
+
+const PopupStyle: Properties = {
+    backgroundColor: 'purple',
+    top: '30%',
+    left: '40%',
+    height: '30%',
+    width: '30%',
+    position: 'absolute',
+    zIndex: 1,
+};
+
+type ServicePopupProps = {
+    state: AppState;
+    dispatch: Dispatch;
+};
+
+export const ServicePopup = (Props: ServicePopupProps) => {
+    const { state, dispatch } = Props;
+    const { serviceSetupIsActive, newService } = state.metaData;
+
+    const [name, setName] = useState('ServiceName');
+    const [port, setPort] = useState(22);
+
+    const addService = (event: MouseEvent) => {
+        event.stopPropagation();
+        const serviceId: AppState['services'][0]['id'] = state.services.length ? state.services.slice(-1)[0].id + 1 : 1;
+        dispatch({ type: 'add-service', serviceToAdd: { ...newService, name, port, id: serviceId } });
+    };
+
+    const closePopup = (event: MouseEvent) => {
+        event.stopPropagation();
+        dispatch({ type: 'cancel-add-service' });
+    };
+
+    const changeName = (event: ChangeEvent<HTMLInputElement>) => {
+        event.stopPropagation();
+        setName(event.target.value);
+    };
+
+    const changePort = (event: ChangeEvent<HTMLInputElement>) => {
+        event.stopPropagation();
+        setPort(parseInt(event.target.value));
+    };
+
+    return (
+        <div style={{ ...PopupStyle, visibility: serviceSetupIsActive ? 'visible' : 'hidden' }}>
+            Name: <input type={'text'} value={name} onChange={changeName} />
+            <br />
+            Port: <input type={'number'} value={port} onChange={changePort} />
+            <br />
+            <button onClick={addService}>Add Service</button>
+            <button onClick={closePopup}>Exit</button>
         </div>
     );
 };
