+++ src/components/index.ts
@@ -2,6 +2,5 @@ export * from './App';
 export * from './Box';
 export * from './Connection';
 export * from './Menu';
-export * from './Popup';
 export * from './Service';
 export * from './Tunnel';
